Change AppStore for MenuStore
